Removed stray debugging info from ACL editor success notification; added ability for guests to set language on URI (?lang=eng); added html_in_pages ACL type and separated from php_in_pages so HTML can be embedded but not PHP; rewote portions of the path manager to better abstract URL input; added Zend Framework into list of BSD-licensed libraries; localized some remaining strings; got the migration script working, but just barely; fixed display bug in Special:Contributions; localized Main Page button in admin panel

+++ includes/lang.php
@@ -360,26 +360,33 @@ $lang_cache = ');
     }
     
     $delete_list = implode(" OR\n  ", $delete_list);
-    $sql = "DELETE FROM " . table_prefix . "language_strings WHERE $delete_list;";
     
-    // Free some memory
-    unset($delete_list);
+    if ( !empty($delete_list) )
+    {
+      $sql = "DELETE FROM " . table_prefix . "language_strings WHERE $delete_list;";
+      
+      // Free some memory
+      unset($delete_list);
+      
+      // Run the query
+      $q = $db->sql_query($sql);
+      if ( !$q )
+        $db->_die('lang.php - couldn\'t kill off them old strings');
+    }
     
-    // Run the query
-    $q = $db->sql_query($sql);
-    if ( !$q )
-      $db->_die('lang.php - couldn\'t kill off them old strings');
-    
-    $insert_list = implode(",\n  ", $insert_list);
-    $sql = "INSERT INTO " . table_prefix . "language_strings(lang_id, string_category, string_name, string_content) VALUES\n  $insert_list;";
-    
-    // Free some memory
-    unset($insert_list);
-    
-    // Run the query
-    $q = $db->sql_query($sql);
-    if ( !$q )
-      $db->_die('lang.php - couldn\'t insert strings in import()');
+    if ( !empty($insert_list) )
+    {
+      $insert_list = implode(",\n  ", $insert_list);
+      $sql = "INSERT INTO " . table_prefix . "language_strings(lang_id, string_category, string_name, string_content) VALUES\n  $insert_list;";
+      
+      // Free some memory
+      unset($insert_list);
+      
+      // Run the query
+      $q = $db->sql_query($sql);
+      if ( !$q )
+        $db->_die('lang.php - couldn\'t insert strings in import()');
+    }
     
     // YAY! done!
     // This will regenerate the cache file if possible.
